fix & finish auth
