+++ packages/app/server/api/auth/logout.get.ts
@@ -1,8 +1,5 @@
 export default defineEventHandler(async (event) => {
   const session = await useAuthSession(event);
   await session.clear();
-
-  return {
-    ok: true,
-  };
+  return sendRedirect(event, '/');
 });
